Updated control.lua
Added a validity check on entity variable on line 205 to prevent from crashing in certain circumstances, of which I am not entirely sure. Most probably the fact that I was trying to place a roboport on a space platform (Space Exploration + place roboports in empty space) which was halfway constructed.
Note that the roboport may not be upgraded according to new technologies when that occurs, but then you get to continue playing. You can destroy and recreate the roboport (if the problem does not reoccur) to ensure that new technologies have an impact.